% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.templateList}
\alias{.templateList}
\title{Create template list}
\usage{
.templateList()
}
\value{
A list.
}
\description{
\code{.templateList} turns the template file in
\url{https://github.com/waldronlab/curatedMetagenomicDataCuration/blob/master/inst/extdata/template.csv}
into a list.
}
\details{
This code is base on:
\url{https://github.com/waldronlab/curatedMetagenomicDataCuration/blob/master/vignettes/curatedMetagenomicDataCuration.Rmd}
}
\keyword{internal}
