% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myApp.R
\name{myApp}
\alias{myApp}
\title{Run App}
\usage{
myApp()
}
\value{
An instance of a shinyApp.
}
\description{
\code{myApp} runs a shinyApp for the \code{get_metadata} and
\code{checkCuration} functions of the curatedMetagenomicDataCuration
package.
}
\examples{

\dontrun{

myApp()

}

}
